refactor(bucketsync): add several capabilities

+++ mobile/bluetooth/data/src/main/kotlin/com/matejdro/catapult/bluetooth/WatchSyncerImpl.kt
@@ -49,7 +49,7 @@ class WatchSyncerImpl(
    }
 
    override suspend fun deleteDirectory(id: Int) {
-      bucketSyncRepository.deleteBucket(id.toUShort())
+      bucketSyncRepository.deleteBucket(id.toUByte())
    }
 }
